# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:1337",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/customer/update": {
      "post": {
        "tags": [
          "customer"
        ],
        "summary": "Update customer",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateCustomer",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created customer object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Customer"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/customer/delete": {
      "delete": {
        "tags": [
          "customer"
        ],
        "summary": "Delete customer",
        "description": "",
        "operationId": "deleteCustomer",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "customer id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
    "/documents/find": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "Get all Documents",
        "description": "",
        "operationId": "getDocuments",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Documents"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Documents not found"
          }
        }
      }
    },
    "/documents/update": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "Update Documents",
        "description": "This can only be done by the logged in user.",
        "operationId": "updateDocuments",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created Documents object",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Documents"
            }
          }
        ],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/documents/delete": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "Delete Documents",
        "description": "",
        "operationId": "deleteDocuments",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Documents id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "string"
            },
            "headers": {
              "X-Rate-Limit": {
                "type": "integer",
                "format": "int32",
                "description": "calls per hour allowed by the user"
              },
              "X-Expires-After": {
                "type": "string",
                "format": "date-time",
                "description": "date in UTC when token expires"
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Customer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "mobile": {
          "type": "string"
        },
        "cnic": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "customer Status"
        }
      },
      "xml": {
        "name": "Customer"
      }
    },
    "Documents": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "file_name": {
          "type": "string"
        },
        "file_path": {
          "type": "string"
        },
        "customer_id": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Documents Status"
        }
      },
      "xml": {
        "name": "Documents"
      }
    },
...
  },
...
}